Classification, Increased sampling (with time-domain)
In P300Speller, additions were made so that there would be a classification button that would run 5 sets of 25 trials in order to get data. 
In the DataProcessing_User file, changes were made (still far from complete) so that data from the graph (500 samples, true time-domain) can be extracted after some period of time. Because of the possibility, I added it and kept the previous stuff. The data from the graph will be extracted in a separate thread and then a case statement at the end will decide whether or not it will analyze the 48 samples we had prior, store the new time-domain samples, or classify them. Storing and Loading data are still incomplete, but the general idea for classification (that I was thinking of) was that we pass in each data point into the classifier. If it is the target, then we will  return true. If not, then it will return false. If it's true, then the "hit_count" (similar to what we had before) will increment. If false, then another "hit_count" variable pertaining to non-target detections will be incremented.  It will do this for some amount of samples until finishes. Additionally, code in the DataProcessing_User was cleaned up a bit, but the P300Speller is still pretty messy.